exercise 06_01_arrays: replace custom function by From<&Weekday> for usize implementation
In the previous parts of the tutorial, the reader is taught about idiomatic ways to do fallible and infallible conversions between types. Later, the exercise 06_01_arrays gives an opportunity to use that knowledge by requiring a conversion between an enum reference `&Weekday` into an usize.

As of now, the `solutions` branch implements a custom `weekday2index` method to do the conversion. It misses an opportunity to show the learners an application of what they learned in the previous chapters about the `From` trait.

Update the solution of exercise 06_01_arrays to use an implementation of the `From` trait to convert `&Weekday` references to `usize` instead of using a `weekday2index` function.